requirements.txt
asgiref==3.5.2
Django==3.2.4
djangorestframework==3.12.4
pytz==2022.1
sqlparse==0.4.2
psycopg2-binary>2.8,<2.9
drf-spectacular==0.15